+++ src/Operators/RIV.cpp
@@ -89,9 +89,9 @@ Parameters : exact results, prediction results
 Inputs     : const vector<float>, const vector<float>
 Outputs    : double
 */
-double RIV::apply(const std::vector<float> &exact,
+float RIV::apply(const std::vector<float> &exact,
                              const std::vector<float> &prediction) {
-  double res = 0;
+  float res = 0;
   float size = prediction.size();
 
   for (unsigned long int i = 0; i < size; ++i) {
